core/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
opencv-python==4.10.0.84
Pillow==12.1.1
scikit-learn==1.7.2
numpy>=2.0.0
python-dotenv==1.1.0
groq==0.25.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
loguru==0.7.3